Colors change from double to uint8_t

+++ sources/libpassion/include/modules/graphics/ps_color.h
@@ -1,11 +1,12 @@
 #ifndef LIBPASSION_INCLUDE_MODULES_GRAPHICS_PS_COLOR_H_
 #define LIBPASSION_INCLUDE_MODULES_GRAPHICS_PS_COLOR_H_
+#include <stdint.h>
 
 struct ps_color {
-        double red;
-        double green;
-        double blue;
-        double alpha;
+        uint8_t red;
+        uint8_t green;
+        uint8_t blue;
+        uint8_t alpha;
 };
 
 #endif
